requirements.txt
google-api-python-client==2.70.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
openai
gtts